# Optimize the performance of the Clone method of preFilterState